trying to fix abs() unittest

+++ modules/c++/types/include/types/complex_short.h
@@ -78,7 +78,7 @@ static_assert(sizeof(complex_short) == sizeof(float), "sizeof(complex_short) != 
 // https://en.cppreference.com/w/cpp/numeric/complex/abs
 inline auto abs(const complex_short& z)
 {
-    const std::complex<float> z_(z.real(), z.imag());
+    const std::complex<long double> z_(z.real(), z.imag());
     const auto result = abs(z_);
     return gsl::narrow_cast<complex_short::value_type>(result);
 }
